pyproject.toml
[project]
name = "patrol-subnet"
dynamic = ["version"]
description = "Patrol Subnet"
requires-python = ">=3.11"
dependencies = [
    "bittensor==9.1.0",
    "aiochclient[aiohttp]",
    "aiohttp",
    "sqlalchemy",
    "aiosqlite",
    "psycopg2-binary",
    "asyncpg",
    "alembic",
    "python-json-logger",
    "pymongo",
    "python-dotenv"
]

[project.optional-dependencies]
test = ["pytest", "pytest_asyncio", "pytest-mock", "httpx", "flake8"]

[build-system]
requires = ["setuptools>=57", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope="function"
addopts = "--capture=no"
log_cli = false
#log_cli_level = "DEBUG"
#log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
#log_cli_date_format = "%Y-%m-%d %H:%M:%S"

pythonpath = ['src/']
testpaths = ['test']
#addopts = ['--strict-config', '-ra']

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"
